Fixes #31922 - Don't audit denials for rpm_script_tmp_t
https://bugzilla.redhat.com/show_bug.cgi?id=1651826 is the root cause
for this. It was closed as WONTFIX for rh-ruby25 and RHEL8 so we'll
likely see these for a while.

Having AVCs in audit logging that are "normal" means you can't easily
distinguish real and "accepted". These AVCs should not be logged.

In fc3eb992a285bc701e8327815c43761a094b2a67 there was already a
workaround for this but in the cron context. This commit removes those
and changes it to the same dontaudit statements as for rpm_script_tmp_t.